refactor: deduplicate Discord configuration check
Extract isDiscordLoginEnabled() into discord-utils.ts to avoid
duplicated logic between user.ts and auth/index.ts. Prevents
maintenance issues from configuration checks diverging.
